Moves all integration tests to an integration folder and provides a new tox env to run only those tests
