+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-norecursedirs = .git build virtualenv teuthology.egg-info .tox
+norecursedirs = .git build virtualenv teuthology.egg-info .tox */integration
