+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27, flake8
+envlist = py27, py27-integration, flake8
 
 [testenv:py27]
 sitepackages=True
@@ -13,6 +13,20 @@ deps=
   coverage==3.7.1
 
 commands=py.test --cov=teuthology --cov-report=term -v {posargs:teuthology scripts}
+
+[testenv:py27-integration]
+sitepackages=True
+deps=
+  -r{toxinidir}/requirements.txt
+  pytest
+  mock
+  fudge
+  nose
+  pytest-cov==1.6
+  coverage==3.7.1
+
+commands=py.test --cov=teuthology --cov-report=term -v {posargs:teuthology/test/integration teuthology/orchestra/test/integration}
+basepython=python2.7
 
 [testenv:flake8]
 deps=
